+++ main/pieces.py
@@ -1,4 +1,8 @@
 class Piece:
+
+    #Piece class is the base class for all the pieces
+    #They get their color and str method from here :)
+
     def __init__(self, color):
         self.__color__ = color
 
@@ -7,9 +11,3 @@ class Piece:
 
     def get_color(self):
         return self.__color__
-    
-    def valid_moves(self, from_row, from_col, to_row, to_col):
-        ...
-    
-    def valid_moves_pawn(self, from_row, from_col, to_row, to_col):
-        ...